feature/migrate-generic-specialization-to-generator (#34)
* update dependency version

* add package reference to GenericSpecializationGenerator

* migrates generic specialization to generator for ScalarOp.BitOperations

* migrates generic specialization to generator for ScalarOp.Sign

* migrate ScalarOp.FusedMultiplyAdd

* migrate ScalarOp.[Abs, Min, Max, Clamp]

* migrate ScalarOp._Math.Real

* migrates ScalarOp.[DivideFloor, DivideCeiling, DivRem]

* moves reference source of GenericSpecializationGenerator from NuGet package to local subrepos

* migrates generic specialization to generator for binary operators in ScalarOp

* migrates ScalarOp.[Equals, Compare]

* migrates ScalarOp.[ShiftLeft, ShiftRight]

* migrates ScalarOp.[UnaryPlus, UnaryMinus, Not, Complement]

* migrates ScalarOp.[AddSaturate, SubtractSaturate]

* migrates VectorOp.[*]

* add .gitmodules

* update GitHub Actions script

* migrate slnx

* fix workflow

* fix subrepo version

+++ src/SmartVectorDotNet.Test/ScalarOpTest/ScalarOpTest._BitOperations.CountTrailingZeros.cs
@@ -24,7 +24,7 @@ public partial class ScalarOpTest
         }
 
         static void testCore(uint x)
-            => Assert.Equal(ctzReference(x), ScalarOp.CountTrailingZeros(x));
+            => Assert.Equal(ctzReference(x), (int)ScalarOp.CountTrailingZeros(x));
 
         static int ctzReference(uint x)
         {
@@ -59,7 +59,7 @@ public partial class ScalarOpTest
         }
 
         static void testCore(ulong x)
-            => Assert.Equal(ctzReference(x), ScalarOp.CountTrailingZeros(x));
+            => Assert.Equal(ctzReference(x), (int)ScalarOp.CountTrailingZeros(x));
 
         static int ctzReference(ulong x)
         {
